+++ tests/testthat/test-connection.R
@@ -62,3 +62,15 @@ test_that("vroom works with file connections and quoted fields", {
   })
   expect_equal(x, tibble::tibble(a = c(1, 4), b = c(2, 5), c = c(3, 6)))
 })
+
+test_that("vroom works with windows newlines and a connection size that lies directly on the newline", {
+  tf <- tempfile()
+  on.exit(unlink(tf))
+
+  writeChar("1,2\r\na,bbb\r\ne,f\r\n", "out", eos = NULL)
+
+  withr::with_envvar(c("VROOM_CONNECTION_SIZE" = 12), {
+    x <- vroom(file("out"), col_types = "cc")
+  })
+  expect_equal(x[[1]], c("a", "e"))
+})